edit: exercise/square-root - missing size_t

+++ square-root/square_root.c
@@ -1,9 +1,10 @@
 #include "square_root.h"
+#include <stddef.h>
 
 unsigned int square_root(unsigned int const num)
 {
   unsigned int arr[num / 2];
-  for (unsigned int i = 0; i <= num / 2; i++)
+  for (size_t i = 0; i <= num / 2; i++)
     arr[i] = i + 1;
 
   unsigned int left = 0, right = (num - 1) / 2;
